[wrangler] Move pages/ commands to src/commands/pages/ (Phase 2e)
