+++ packages/wrangler/src/__tests__/pages/deployment-list.test.ts
@@ -6,7 +6,7 @@ import { mockAccountId, mockApiToken } from "./../helpers/mock-account-id";
 import { msw } from "./../helpers/msw";
 import { runInTempDir } from "./../helpers/run-in-tmp";
 import { runWrangler } from "./../helpers/run-wrangler";
-import type { Deployment } from "./../../pages/types";
+import type { Deployment } from "../../commands/pages/types";
 
 describe("pages deployment list", () => {
 	runInTempDir();
