+++ packages/wrangler/src/__tests__/pages/filepath-routing.test.ts
@@ -3,10 +3,10 @@ import { describe, expect, it, test } from "vitest";
 import {
 	compareRoutes,
 	generateConfigFromFileTree,
-} from "../../pages/functions/filepath-routing";
+} from "../../commands/pages/functions/filepath-routing";
 import { toUrlPath } from "../../paths";
 import { runInTempDir } from "../helpers/run-in-tmp";
-import type { HTTPMethod, RouteConfig } from "../../pages/functions/routes";
+import type { HTTPMethod, RouteConfig } from "../../commands/pages/functions/routes";
 import type { UrlPath } from "../../paths";
 
 describe("filepath-routing", () => {
