+++ packages/wrangler/src/assets.ts
@@ -21,10 +21,10 @@ import PQueue from "p-queue";
 import prettyBytes from "pretty-bytes";
 import { FormData } from "undici";
 import { fetchResult } from "./cfetch";
+import { hashFile } from "./commands/pages/hash";
+import { isJwtExpired } from "./commands/pages/upload";
 import { formatTime } from "./deploy/deploy";
 import { logger, LOGGER_LEVELS } from "./logger";
-import { hashFile } from "./pages/hash";
-import { isJwtExpired } from "./pages/upload";
 import { getBasePath } from "./paths";
 import { dedent } from "./utils/dedent";
 import type { StartDevWorkerOptions } from "./api";
