+++ packages/wrangler/src/commands/pages/download-config.ts
@@ -8,13 +8,13 @@ import {
 import chalk from "chalk";
 import { supportedCompatibilityDate } from "miniflare";
 import TOML from "smol-toml";
-import { fetchResult } from "../cfetch";
-import { getConfigCache } from "../config-cache";
-import { createCommand } from "../core/create-command";
-import { confirm } from "../dialogs";
-import { logger } from "../logger";
-import * as metrics from "../metrics";
-import { requireAuth } from "../user";
+import { fetchResult } from "../../cfetch";
+import { getConfigCache } from "../../config-cache";
+import { createCommand } from "../../core/create-command";
+import { confirm } from "../../dialogs";
+import { logger } from "../../logger";
+import * as metrics from "../../metrics";
+import { requireAuth } from "../../user";
 import { PAGES_CONFIG_CACHE_FILENAME } from "./constants";
 import type { PagesConfigCache } from "./types";
 import type { Project } from "@cloudflare/types";
