+++ packages/wrangler/src/commands/pages/functions/buildPlugin.ts
@@ -1,11 +1,11 @@
 import { access, lstat } from "node:fs/promises";
 import { relative, resolve } from "node:path";
-import { bundleWorker } from "../../deployment-bundle/bundle";
-import { createModuleCollector } from "../../deployment-bundle/module-collection";
-import { getBasePath } from "../../paths";
+import { bundleWorker } from "../../../deployment-bundle/bundle";
+import { createModuleCollector } from "../../../deployment-bundle/module-collection";
+import { getBasePath } from "../../../paths";
 import { getPagesProjectRoot } from "../utils";
 import { buildNotifierPlugin } from "./buildWorker";
-import type { Entry } from "../../deployment-bundle/entry";
+import type { Entry } from "../../../deployment-bundle/entry";
 import type { Options as WorkerOptions } from "./buildWorker";
 
 type Options = Omit<
