+++ packages/wrangler/src/commands/pages/functions/buildWorker.ts
@@ -3,17 +3,17 @@ import { access, cp, lstat, rm } from "node:fs/promises";
 import { join, resolve } from "node:path";
 import { FatalError } from "@cloudflare/workers-utils";
 import { build as esBuild } from "esbuild";
-import { bundleWorker } from "../../deployment-bundle/bundle";
-import { findAdditionalModules } from "../../deployment-bundle/find-additional-modules";
+import { bundleWorker } from "../../../deployment-bundle/bundle";
+import { findAdditionalModules } from "../../../deployment-bundle/find-additional-modules";
 import {
 	createModuleCollector,
 	noopModuleCollector,
-} from "../../deployment-bundle/module-collection";
-import { logBuildFailure, logger } from "../../logger";
-import { getBasePath } from "../../paths";
+} from "../../../deployment-bundle/module-collection";
+import { logBuildFailure, logger } from "../../../logger";
+import { getBasePath } from "../../../paths";
 import { getPagesProjectRoot, getPagesTmpDir } from "../utils";
-import type { BundleResult } from "../../deployment-bundle/bundle";
-import type { Entry } from "../../deployment-bundle/entry";
+import type { BundleResult } from "../../../deployment-bundle/bundle";
+import type { Entry } from "../../../deployment-bundle/entry";
 import type { CfModule } from "@cloudflare/workers-utils";
 import type { Plugin } from "esbuild";
 import type { NodeJSCompatMode } from "miniflare";
