+++ packages/wrangler/src/commands/pages/functions/filepath-routing.ts
@@ -2,9 +2,9 @@ import fs from "node:fs/promises";
 import path from "node:path";
 import { FatalError } from "@cloudflare/workers-utils";
 import { build } from "esbuild";
-import { toUrlPath } from "../../paths";
+import { toUrlPath } from "../../../paths";
 import { FunctionsBuildError } from "../errors";
-import type { UrlPath } from "../../paths";
+import type { UrlPath } from "../../../paths";
 import type { HTTPMethod, RouteConfig } from "./routes";
 
 export async function generateConfigFromFileTree({
