+++ packages/wrangler/src/commands/pages/functions/identifiers.ts
