+++ packages/wrangler/src/commands/pages/functions/routes-consolidation.ts
