+++ packages/wrangler/src/commands/pages/functions/routes.ts
@@ -2,7 +2,7 @@ import fs from "node:fs/promises";
 import path from "node:path";
 import { UserError } from "@cloudflare/workers-utils";
 import { isValidIdentifier, normalizeIdentifier } from "./identifiers";
-import type { UrlPath } from "../../paths";
+import type { UrlPath } from "../../../paths";
 
 export type HTTPMethod =
 	| "HEAD"
