+++ packages/wrangler/src/commands/pages/functions/tsconfig.json
