+++ packages/wrangler/src/commands/pages/hash.ts
