+++ packages/wrangler/src/commands/pages/index.ts
@@ -1,4 +1,4 @@
-import { createNamespace } from "../core/create-command";
+import { createNamespace } from "../../core/create-command";
 import { CLEANUP } from "./utils";
 
 export const pagesNamespace = createNamespace({
