+++ packages/wrangler/src/commands/pages/secret/index.ts
@@ -5,17 +5,17 @@ import {
 	findWranglerConfig,
 } from "@cloudflare/workers-utils";
 import chalk from "chalk";
-import { fetchResult } from "../../cfetch";
-import { parseBulkInputToObject } from "../../commands/secret";
-import { readPagesConfig } from "../../config";
-import { getConfigCache } from "../../config-cache";
-import { createCommand, createNamespace } from "../../core/create-command";
-import { confirm, prompt } from "../../dialogs";
-import isInteractive from "../../is-interactive";
-import { logger } from "../../logger";
-import * as metrics from "../../metrics";
-import { requireAuth } from "../../user";
-import { readFromStdin, trimTrailingWhitespace } from "../../utils/std";
+import { fetchResult } from "../../../cfetch";
+import { readPagesConfig } from "../../../config";
+import { getConfigCache } from "../../../config-cache";
+import { createCommand, createNamespace } from "../../../core/create-command";
+import { confirm, prompt } from "../../../dialogs";
+import isInteractive from "../../../is-interactive";
+import { logger } from "../../../logger";
+import * as metrics from "../../../metrics";
+import { requireAuth } from "../../../user";
+import { readFromStdin, trimTrailingWhitespace } from "../../../utils/std";
+import { parseBulkInputToObject } from "../../secret";
 import { PAGES_CONFIG_CACHE_FILENAME } from "../constants";
 import { EXIT_CODE_INVALID_PAGES_CONFIG } from "../errors";
 import type { PagesProject } from "../download-config";
