+++ packages/wrangler/src/commands/pages/utils.ts
@@ -1,7 +1,7 @@
 import path from "node:path";
 import { findUpSync } from "find-up";
-import { getWranglerTmpDir } from "../paths";
-import type { BundleResult } from "../deployment-bundle/bundle";
+import { getWranglerTmpDir } from "../../paths";
+import type { BundleResult } from "../../deployment-bundle/bundle";
 
 export const RUNNING_BUILDERS: BundleResult[] = [];
