+++ packages/wrangler/src/miniflare-cli/assets.ts
@@ -7,8 +7,8 @@ import { watch } from "chokidar";
 import { getType } from "mime";
 import { fetch, Request, Response } from "miniflare";
 import { Dispatcher, getGlobalDispatcher } from "undici";
+import { hashFile } from "../commands/pages/hash";
 import { logger } from "../logger";
-import { hashFile } from "../pages/hash";
 import type { Logger } from "../logger";
 import type { Metadata } from "@cloudflare/pages-shared/asset-server/metadata";
 import type {
